more refactoring on Form and movieResults to make App cleaner
