+++ src/components/MovieResults.js
@@ -1,7 +1,7 @@
 const MovieResults = ({ movies }) => {
   return (
     <ul>
-      {movies.map((individualMovie) => {
+      {movies?.map((individualMovie) => {
         return (
           <li key={individualMovie.id}>
             <p>{individualMovie.title}</p>
